Renamed test files and removed some duplicate code
