+++ tlstest/certrequireandverify/certrequireandverify_test.go
@@ -1,4 +1,4 @@
-package certrequest
+package certrequireandverify
 
 import (
 	"context"
@@ -10,16 +10,23 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestAccessOverPlaintext(t *testing.T) {
-	dg := z.DgraphClient(":9180")
-	err := dg.Alter(context.Background(), &api.Operation{DropAll: true})
-	require.Error(t, err, "The authentication handshake should have failed")
-}
-
-func TestAccessWithCaCert(t *testing.T) {
+func TestAccessWithoutClientCert(t *testing.T) {
 	conf := viper.New()
 	conf.Set("tls_cacert", "./tls/ca.crt")
 	conf.Set("tls_server_name", "node")
+
+	dg, err := z.DgraphClientWithCerts(":9180", conf)
+	require.NoError(t, err, "Unable to get dgraph client: %v", err)
+	err = dg.Alter(context.Background(), &api.Operation{DropAll: true})
+	require.Error(t, err, "The authentication handshake should have failed")
+}
+
+func TestAccessWithClientCert(t *testing.T) {
+	conf := viper.New()
+	conf.Set("tls_cacert", "./tls/ca.crt")
+	conf.Set("tls_server_name", "node")
+	conf.Set("tls_cert", "./tls/client.acl.crt")
+	conf.Set("tls_key", "./tls/client.acl.key")
 
 	dg, err := z.DgraphClientWithCerts(":9180", conf)
 	require.NoError(t, err, "Unable to get dgraph client: %v", err)
